Centralize cell geometry and distinguish unsupported from malformed selection
