+++ packages/slam-rs/crates/slam-rs/src/gpu/kernels.rs
@@ -1723,9 +1723,9 @@ const SELECT_EDGE: f32 = crate::frontend::detect::EDGE_THRESHOLD;
 // and `detect_keypoints_with_cells` takes them apart, and a shift that drifted
 // by one would move every corner without failing to compile.
 /// Where the packed key keeps `255 - score`.
-const KEY_SCORE_SHIFT: u32 = crate::frontend::detect::KEY_SCORE_SHIFT;
+const KEY_SCORE_SHIFT: u32 = crate::frontend::cell::KEY_SCORE_SHIFT;
 /// Where the packed key keeps the row.
-const KEY_ROW_SHIFT: u32 = crate::frontend::detect::KEY_ROW_SHIFT;
+const KEY_ROW_SHIFT: u32 = crate::frontend::cell::KEY_ROW_SHIFT;
 /// `NO_CELL_WINNER`, which the kernel spells as a literal below because the
 /// `#[cube]` macro keeps a *named* integer constant comptime and the local it
 /// initialises is then assigned from the runtime reduction. This is what pins
